Move ColumnConfigurations to columns config.
